% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_function.R
\name{eval}
\alias{eval}
\title{Evaluate Performance Metrics for Differential Abundance Analysis}
\usage{
eval(Res, isDA = Res$isDA, new_isDA = Res$new_ind)
}
\arguments{
\item{Res}{A list containing results from the original dataset, including true and newly identified DA statuses.}

\item{isDA}{Logical vector. The ground truth for differential abundance status in the data (default: \code{Res$isDA}).}

\item{new_isDA}{Character vector. The newly classified differential abundance status (default: \code{Res$new_ind}).}
}
\value{
A list containing:
\item{FDR}{False Discovery Rate, computed as FP / (FP + TP).}
\item{Sensitivity}{True Positive Rate (TPR), computed as TP / (TP + FN).}
\item{Specificity}{True Negative Rate (TNR), computed as TN / (TN + FP).}
\item{Type_1_error}{Type I Error Rate, computed as FP / (FP + TN).}
\item{Power}{Statistical power, computed as 1 - beta, where beta is FN / (FN + TP).}
}
\description{
This function computes key performance metrics, including False Discovery Rate (FDR), Sensitivity, Specificity, Type I Error Rate, and Power,
based on the classification of taxa as differentially abundant (DA) or not.
}
