% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_function.R
\name{Normalize}
\alias{Normalize}
\title{Normalize two samples using percentile-based scaling/normalization}
\usage{
Normalize(sample1, sample2)
}
\arguments{
\item{sample1}{A numeric vector representing the first sample of data.}

\item{sample2}{A numeric vector representing the second sample of data.}
}
\value{
A list with two elements:
\describe{
\item{sample1}{The normalized version of the first sample.}
\item{sample2}{The normalized version of the second sample.}
}
}
\description{
This function normalizes two input samples by performing a percentile-based scaling. It pools the two samples together,
calculates the 0th and 99th percentiles of the pooled data, and then scales the values within each sample.
An additional adjustment is made based on the median values of the two samples.
}
\details{
The function pools the two samples to calculate quantiles (0th and 99th) and uses the range between these
quantiles as the normalization factor. If there are many zeros, the function defaults to using half the maximum value of the pooled data
as the normalization factor. The normalized samples are then adjusted by the average of the medians of the two input samples.
}
\examples{
sample1 <- c(1, 2, 3, 4, 5)
sample2 <- c(2, 3, 4, 5, 6)
result <- Normalize(sample1, sample2)
result$sample1
result$sample2

}
