% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_stat_function.R
\name{Test_Statistic}
\alias{Test_Statistic}
\title{Compute Test Statistic for Differential Abundance Analysis}
\usage{
Test_Statistic(org_otu, Train_parest, n1, n2, B)
}
\arguments{
\item{org_otu}{A matrix representing the OTU table with samples in rows (OTU tavle from original dataset)}

\item{Train_parest}{A vector of estimated parameters (scores) used for computing the test statistic.}

\item{n1}{The number of samples in group 1 (default: half of the dataset).}

\item{n2}{The number of samples in group 2 (default: half of the dataset).}

\item{B}{Number of permutations used for significance testing (default: 100).}
}
\value{
A vector of computed test statistics for each taxon.
}
\description{
This function calculates a test statistic for each taxon using a PLS estimates/scores from the trained model.
It sorts samples within each group and aligns them according to their mean abundance before computing the test statistic.
}
