% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adaptive_function.R
\name{ADATEST}
\alias{ADATEST}
\title{Adaptive Test for Differential Abundance Analysis}
\usage{
ADATEST(physeq, t0 = 1, t1 = 1.5, t2 = +Inf, B = 1000)
}
\arguments{
\item{physeq}{A phyloseq object containing OTU data.}

\item{t0}{Lower bound for data transformation (default: 1).}

\item{t1}{Upper bound for data transformation (default: 1.5).}

\item{t2}{Maximum bound for data transformation (default: +Inf).}

\item{B}{Number of permutations used for significance testing (default: 1000).}
}
\value{
A list containing:
\itemize{
\item \code{Original_results}: Results from the permutation test.
\item \code{Unscaled_data}: List containing training data and original data.
\item \code{Train_results}: Results from training data analysis.
\item \code{Results}: Evaluation metrics including FDR, sensitivity, specificity, and power.
}
}
\description{
This function performs an adaptive differential abundance test using a permutation-based approach.
It applies total sum scaling (TSS), trims the dataset to remove taxa that are not prevalent,
subsets the data to create the Pseudo dataset and finally the training dataset,
performs parameter estimation, conducts a permutation test, and evaluates performance metrics.
}
