% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudo_data_function.R
\name{PseudoData}
\alias{PseudoData}
\title{PseudoData Function}
\usage{
PseudoData(
  simdata_filter,
  group_var = "group",
  group_levels = c("0", "1"),
  t0 = 1,
  t1 = 1.5,
  t2 = +Inf,
  empirical_adjust = TRUE,
  n.taxa0 = NULL
)
}
\arguments{
\item{simdata_filter}{A phyloseq object containing the microbiome data.}

\item{t0}{Lower threshold for log fold change (LFC) classification of pseudo taxa (default: 1).}

\item{t1}{Upper threshold for DA classification (default: 1.5).}

\item{t2}{Maximum threshold for DA classification (default: +Inf).}

\item{empirical_adjust}{Logical; if TRUE, adjusts the pseudo-dataset based on empirical distribution.}

\item{n.taxa0}{Number of taxa in the non-DA group, if known (if default (NULL), it is set dynamically
based on half the total number of taxa.).}
}
\value{
A list containing:
\item{pseudo_lfc}{Log fold change values of pseudo-taxa.}
\item{Train}{Training dataset for downstream analysis.}
}
\description{
This function constructs a pseudo-dataset from a given phyloseq object by generating
artificial pseudo-taxa from the group with the largest sample size. It is designed to
support differential abundance (DA) analysis while maintaining valid inference without bias.
}
\details{
\enumerate{
\item \strong{Pseudo-Dataset Construction}: The function selects the group with the largest
sample size and constructs pseudo-taxa by pairing taxa within this group.
\item \strong{Empirical Adjustment}: If \code{empirical_adjust = TRUE}, it downsamples data to match
the original group distribution.
\item \strong{Differential Abundance Classification}: Log fold changes (LFCs) are computed, and
pseudo-taxa are categorized into three subsets based on thresholds \code{t0}, \code{t1}, and \code{t2}:
\itemize{
\item \code{I_0}: Non-differentially abundant pseudo-taxa.
\item \code{I_1}: Differentially abundant pseudo-taxa.
\item \code{I_B}: Biologically irrelevant or extreme DA pseudo-taxa.
}
\item \strong{Training Dataset Generation}: The final dataset is created by normalizing and
partitioning pseudo-taxa based on LFC.
}

This function is useful for developing DA testing procedures that control false
discovery rates while leveraging synthetic data for statistical power.
}
\examples{
result <- PseudoData(simdata_filter)
pseudo_lfc <- result$pseudo_lfc
training_set <- result$Train

}
